Extract field-focus wiring into a controller; move flash styles out; fix wording
- FieldFocusController (a ReactiveController, lifecycle hooks for listener add/remove) owns the focusin/pointerdown/input/change listeners and field-focus emission, shrinking the over-cap config-entry-form.ts (720 -> 658).
- Move the .field--highlight glow + keyframes into field-highlight.styles.ts so the over-cap config-entry-form.styles.ts stops growing; add it to fieldRendererStyles.
- Reword the getKeyPath test and findFieldLine comment: getKeyPath omits list-item wrappers and slice(1) drops the section key; findFieldLine accepts an optional leading section key.
